chore: use unified sli registry in dashboards
In https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2602
we've added a recording rule registry that includes all of the metrics
used in SLIs with all of their significant labels.

This included a rename of a metric, which handily did not clash with
the Thanos rules.

Now that we've switched Grafana to use Mimir as a default datasource,
we can use this registry to resolve wich metrics to display.