chore: make prometheus rule group generator respect config
This makes the `prometheus-service-group-generator` respect the
`gitlab-metrics-config`.

When doing this, we make sure that the GET-hybrid continues to use the
NullRegistry, even when we switch the default over to use the unified
registry.

This also configures the service-key-metric-rule-files that we use for
Prometheus+Thanos over to using the selective registry so we don't
change the recording rules there when switching the default.