fix: make combined metrics respect configured config
Before this, combined metrics would always rely on the default
configuration.

In our Mimir setup, this meant that for these SLIs we would rely on
high cardinality source metrics, while we had recording rules
available.

For the GET-hybrid, this meant that for alerts we were generating a
query that was using recording rules that weren't there.