fix(alert): make Gitaly Version Mismatch non-paging
What
---
Update the `Gitaly Version Mismatch` alert from a paging alert to an
issue alert.

Why
---
Running multiple versions of Gitaly is not a problem for our users.
However in the past running multiple versions showed we had something
in our tooling broken like chef, or deployer.
Which is why we are making this alert as an async alert and not
completely deleting it.

Previous incidents:
- <https://gitlab.com/gitlab-com/gl-infra/production/-/issues/18089>: Chef
  broken
- <https://gitlab.com/gitlab-com/gl-infra/production/-/issues/7897>:
  The Gitaly process was not terminated due to the Gitaly bug.
- <https://gitlab.com/gitlab-com/gl-infra/production/-/issues/6427> :
  The Gitaly process was not terminated due to the Gitaly bug.

Signed-off-by: Steve Xuereb <sxuereb@gitlab.com>