feat: SLI status panels use confidence for status
This updates the status panels to rely on confidence interval values
instead of direct raw value when calculating the status of an SLI.

Also, some more information about the burn rates is included in the
status panel (whether they are degraded, recovering, or deteriorating).

This may help in an incident to understand when things are getting
better or worse.