fix: fix application SLI violations dashboard for mimir
This removes the `monitor="global"` selector when used on the
`gitlab:feature_category:stage_group:mappings` metric. As in Mimir
this metric does not have this label.

Removing the selector should not break anything when using Thanos, as
the recording should be the same in Prometheus and Thanos in this
environment.

This is something we intend to do anyway in
https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2902