refactor: defaultPrometheusDatasource derived from tenants
Previously, a service definition needs to define `tenants` and
`defaultPrometheusDatasource` where the latter is used for the default
PROMETHEUS_DS in dashboards.

This refactors GitLab.com services to use the tenant ID from the first
element of the `tenants` attribute as the default data source, defined
by default in the new `defaultTenant` attribute.

The `defaultTenant` will be used for determining the tenant to query for that
service for Tamland https://gitlab.com/gitlab-com/gl-infra/tamland/-/issues/183