fix: mark sidekiq's urgent-other shard as autoscaling
This shard is autoscaled, as seen in the config:
https://gitlab.com/gitlab-com/gl-infra/k8s-workloads/gitlab-com/blob/039686c78fdd932d5ec05129db69e8073e8f502f/releases/gitlab/values/gprd.yaml.gotmpl#L891-892

The diff is huge because the `shard!=` selector for Sidekiq is applied
to all services for which this saturation point applies. We should
probably change this to be part of the `kubeResources` of a
service. We might have to pick that up in the context of
https://gitlab.com/gitlab-com/gl-infra/mstaff/-/issues/326#note_1955764428

For https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/3613