feat: show the legend on the saturation panel
This adds the legend on the saturation dashboard panel. Having the
legend there makes it easier to see what resource is saturating by
sorting the legend.

Automatically sorting the legend on load would also have been nice,
but that lead me close to
https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2573. And
it didn't fix my timebox.

It also makes it easier to single out a saturation point by clicking
it in the legend.