fix: mark the frontend service as stageless
This marks the frontend service as stageless. As a result, it will get
`stage: 'main'` applied to all alerts that don't have a stage defined
otherwise.

Only the `cnyHttpsServices` and `mainHttpsServices` have a specified
stage label on the SLI, the others will be attributed to `main`.

For https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/3617