fix: add the env GCPScheduledSnapshotsFailed alerts
This adds back the `env` label to the `GCPScheduledSnapshotsFailed`
alerts. Even though we're only selecting one env, we still want the
label to be present on the alert. Including it in the aggregation
makes this alert behave the way it was in Thanos.

For https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/3617