fix: feature category mapping and stage groups availability rules
The feature category mapping and stage group availability rules
previously only existed in gitlab-gprd tenant. These rules are required
to calculate error budgets.

Added a new argument `overrideTenants` for the
`separate-mimir-recording-files` helper to pass in tenants from the
caller instead of following the service's tenant or `defaultTenants`
from the `gitlab-metrics-config`. This allows us to pass in the
`stageGroupTenants` for the rules above, otherwise they are only
generated in `gitlab-gprd` and `gitlab-gstg` (from defaultTenants)