feat: validate teams.yml in the service catalog
This adds validation for `services/teams.yml` that is part of the
service catalog.

With this change we validate that for all teams specified in `teams.yml`
that have a `product_stage_group` specified the stage group refers to
a know group in the `stage-group-mapping.jsonnet`. The latter is
automatically kept up to date.

We also validate the other way around: a stage group should have an
entry in the `teams.yml` file or otherwise be ignored. It should be a
deliberate decision. Otherwise we risk missing alerts for new
groups. For example
https://gitlab.com/gitlab-com/gl-infra/production/-/issues/18191 might
have been noticed sooner if we already had an entry for this group.