runbook: Create gitaly feature status dashboard
Currently, the Gitaly Feature Status dashboard is broken. It is
unmanaged and run the risk of becoming stale. At the same time,
this dashboard has been crucial to identify and debug production
issues. Here we are re-creating this dashboard in the runbooks
repo using jsonnet templates so it becomes a managed dashboard.