fix: show all shards for ci-runner SLIS
Only the `ci_runner_jobs` SLI is sharded, this means that we're using
the `component_shard` aggregation on the dashboard for this SLI and
require a `shard` template for selecting which shard is being viewed.

This adds that shard template to the main dashboard so it can be used
for the SLI panels. By default, all shards will be shown.