fix: Do not generate unnecessary grafana thresholds
Generating a threshold for a grafana panel when there is no actual value results in a failure to render in recent versions of Grafana (see https://gitlab.com/gitlab-com/gl-infra/gitlab-dedicated/incident-management/-/issues/370).

And some services have no toleratedThreshold for some SLIs (noticeably workhorse).

The fix is to not emit that threshold definition if there is no value.