docs: leaving only bypass and how-to rate-limiting info
The rest of the information that was originally in this file has been
moved to the handbook doc in https://gitlab.com/gitlab-com/content-sites/handbook/-/merge_requests/8002

Therefore to remove duplication, this removes the information that no
longer needs to exist in the runbooks.

For now, bypass and "how tos" have been retained, but it's possible some
of this will be moved as a child page of the new handbook page in the
future.