chore: protect rate limiting WIP dashboard
The intention is to replace this with a jsonnet-ified (that's a word) dashboard, but to get us moving we created this dashboard manually to understand the metrics we want to surface (and give an idea of what we're working towards).

We'd like this version to not be deleted until we get a more permanent dashboard in place. :)

More detail: https://gitlab.com/gitlab-com/gl-infra/production-engineering/-/issues/25743