fix: remove broken log link from patroni alert
When the `PatroniLongRunningTransactionDetected` alert fires it includes a link to suggest checking the slowlog for changes in the usual trends. This is currently broken, and can cause confusion for people joining the on-call rotation (like myself). 

The runbook we point to seems to  provide a lot of useful links and steps to investigate, so we can probably remove this link.

Alternatively: It looks like it might intend to link to the [Top PostgreSQL slow log entries Kibana visualisation](https://log.gprd.gitlab.net/app/visualize#/edit/34b49230-7745-11eb-966b-2361593353f9?_g=h@e886e10&_a=h@9aff8b8).

I tried briefly to get that working but need someone with a better unerstanding of jsonnet to help with that if we did want to keep the link.