fix: use only aggregation set labels on MWMBR expressions
This makes sure that we only select for the labels that are present on
the aggregation set. In theory, these should be the only labels
present on the ratios. But in
https://gitlab.com/gitlab-com/gl-infra/gitlab-dedicated/team/-/issues/6378#note_2136974528
we saw that there was an `upscale_source` label present that prevented
us from joining the 5m and the 1h ratios.

This makes sure this doesn't happen again even if we add labels to one
of the ratios but not the other.