fix: remove the `upscale_source` label from aggregation sets
This removes the `upscale_source` label from aggregation sets' rate
metrics.

It also removes the conditional upscaling of aggregation set
transformations in case the direct expression is not possible. This
would happen if we want to add a `6h` window to an aggregation set,
but the source aggregation itself does not have it. We did this
because in the Prometheus world of GitLab.com, some of the
prometheus-local aggregations did not have the 6h burn rate, we'd only
get those in the Thanos aggregations.

We don't need this anymore. For GitLab.com we use
`upscaleLongerBurnrates` to upscale aggregation sets. In that case
they're using their own `1h` rate for upscaling.

For dedicated, we don't need to worry about this, all aggregation sets
have the 6h burn rate. For the component metrics, from source, we're
using the source metrics. For the transformation from component to
service metrics, the required burnRate is present so no need to
upscale.

For https://gitlab.com/gitlab-com/gl-infra/gitlab-dedicated/team/-/issues/6378