feat: Add envOverrides to the metrics catalog service as an option.
This change adds the option for `envOverrides` to be set on service
definitions to override the `env` label that is automatically picked
up based on the configured tenant.

In our case we have metrics for our `http-router` service being
collected in `gitlab-ops`, but due to our alert routing we need to end
up matching on our conventional `gprd` and `gstg` labels.

Without this change we are limited to only matching `env="ops"` inside
of the `gitlab-ops` tenant. Because we're evaluating rules inside of
each tenant the metrics don't exist in `gitlab-gprd` and `gitlab-gstg`
where the rules were previously configured to evaluate.

Only the `http-router` is affected by this change currently, though
for hosted services where we have a central collector we may want to
expand usage of this pattern, e.g. for the `vscode` component of `gitlab-static`.