feat: Create cloudflare-worker archetype
This change moves to create a `cloudflare-worker` service archetype.

We have a few cloudflare workers already deployed. This change makes
it easier to create an entry in the metrics catalog for
cloudflare-worker based services.

This is based off of the work done for the `http-router`
cloudflare-worker service.

---

Also included is a change to allow for addition of logging links to
the tooling links for services, based on those added in the service
catalog, centralising configuration to a single place when in use.