feat: add an SLI to ai-gateway for inferences using other providers
This adds a catchall inference SLI that excludes `anthropic` and
`vertex-ai`, for which we have dedicated SLIs.

We already have `anthropic-chat` in production and `litellm-chat` in
staging in use. We plan on adding another provider in
https://gitlab.com/gitlab-org/gitlab/-/issues/502635

This will make sure we don't entirely miss any newly added providers
going forward.