feat: separate postgres memory saturation monitoring
This replaces the regular `memory` saturation point used for all
services running on VMs, including `patroni`.

For patroni, we now have 2 saturation points:

- `pg_anon_memory`: This tracks how much memory is used for anonymous
  pages (not backed by files). An increase here means that there's
  less memory available for the filesystem cache. The threshold is set
  fairly low because this value should be constant.

- `pg_memory`: Regular memory allocation. This shows memory that is
  explicitly allocated. But with a lower threshold and higher severity
  than we do for the regular `memory` component.

The combination of these saturation points should give us a better
view on how much memory is available on the hosts for filesystem
cache. Having most of the memory available for FS-cache is crucial for
query performance.

Both of these alerts are currently set to page the SRE-oncall. I think
we should address this so the alerts can be routed directly to the
team owners defined in the service
catalog (https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/3941).

Adding these as saturation points will show the aggregated (max across
all the nodes in a service) in the main saturation panel on the top
right of the service dashboards for patroni nodes.

There will be detail panels for both these saturation points collapsed
under "Saturation Details". These panels will include the per-node
breakdown of the utilization, allowing us to identify which node is
using more memory than expected.

For https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/3553
after discussion in
https://gitlab.com/gitlab-com/runbooks/-/merge_requests/7970#note_2153053230