fix: saturation monitoring for total elastic_disk_space
This saturation point is supposed to monitor the total disk space we
have available on ES client nodes. Elastic rebalances within these
nodes as it sees fit, we currently use the
`elastic_single_node_disk_space` to monitor that. This saturation
point is used for monitoring the total space we have available, so we
don't need `host` as `resourceLabel`.

Related to
https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/3974