fix: ai-gateway server_code_completions noisy apdex
What
---
Reduce apdex from 0.98 to 0.95 for `ai-gateway`
`server_code_completions`.

Why
---
In <https://gitlab.com/gitlab-com/gl-infra/production/-/issues/18855>
we are seeing frequent pages for the on-call for this apdex,
that automatically recovers.
In 10 days this paged the [on-call 20 times](https://log.gprd.gitlab.net/app/r/s/OC4z6),
with no action for the on-call to do.

Right now the apdex is set to
[0.98](https://gitlab.com/gitlab-com/runbooks/-/blob/e3002a54b6cc0b65ddd9c6520d439ce9838f06cd/metrics-catalog/services/ai-gateway.jsonnet#L35),
if we look at
[Global SLI Metrics Apdex SLO Analysis](https://dashboards.gitlab.net/d/alerts-component_slo_apdex/alerts3a-global-sli-metrics-apdex-slo-analysis?orgId=1&from=2024-11-19T00:00:00.000Z&to=2024-11-19T23:59:59.000Z&timezone=utc&var-PROMETHEUS_DS=mimir-runway&var-environment=gprd&var-environment-2=gprd&var-type=ai-gateway&var-stage=main&var-component=server_code_completions&var-proposed_slo=0.95&var-confidence=2.3263478740000001)
we see if we set to 0.95 it wouldn't have paged the on-call.

Reference: https://gitlab.com/gitlab-com/gl-infra/production/-/issues/18855