fix: pager noise for HTTP Router CPU saturation
What
---
Set the HTTP Router CPU budget to s3 instead of s2 so it doesn't page the
on-call.

Why
---
We see 4 pages to the on-call in
https://gitlab.com/gitlab-com/gl-infra/production/-/issues/19074#note_2282310000
which aren't actionable. We are already working on this as part of
https://gitlab.com/gitlab-com/gl-infra/capacity-planning-trackers/gitlab-com/-/issues/2026
and have a support ticket with Cloudflare.

In the meantime disable paging the on-call for a non-actionable page.

reference: https://gitlab.com/gitlab-com/gl-infra/production/-/issues/19074
reference: https://gitlab.com/gitlab-com/gl-infra/capacity-planning-trackers/gitlab-com/-/issues/2026