fix(sla-dashboard): use the correct expansion for weighted panel
This correctly expands the query for the SLA from the recording rules.

The difference here is in the fact that we're using a double
`avg_over_time`, once, [simulating the ones from the recording
rules][0].

And once for the total ratio, just like we do on in the system
[availability panel][1].

This also removes the fallback metrics for `slo:observation_status`
since we've fully migrated to Mimir.

Fixes https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/4036

[0]: https://gitlab.com/gitlab-com/runbooks/-/blob/079c1d469dd3ceec05fc13d9f178ce95209b7b8a/mimir-rules/gitlab-gprd/autogenerated-gitlab-gprd-sla-rules.yml#L7-38
[1]: https://gitlab.com/gitlab-com/runbooks/-/blob/079c1d469dd3ceec05fc13d9f178ce95209b7b8a/dashboards/general/slas.dashboard.jsonnet#L52