docs: Add guidance on enabling Application rate limits in dry run mode
What
---

Elaborates on how to enable Application Rate Limits in "Dry Run" mode. 

Why
---
There are times when we are introducing new rate limits to the GitLab Application, and to minimise potential impact, setting these limits in "dry run" mode enables us to monitor potential impact before turning the rate limit on. 

The documentation around this was unclear, so this explicitly outlines the process for doing so. 

Prompted by: https://gitlab.com/gitlab-com/gl-infra/production/-/issues/19084