feat: record availability sla rules for all services
This separates the recording rules for GitLab.com's occurrence based
Availability by service. We needed to do that in order to be able to safely
record these rules for all services rather than just the key services.

As an added bonus, this means that we can now record these for the
right tenant based on the service.

This is part of the discussion in
https://gitlab.com/gitlab-com/gl-infra/observability/team/-/issues/4082