chore: enable new timeseries panel for more service dashboards
This enables the new timeseries panel, replacing graphpanel for more
service dashboards.

Doing it for more of them, that other people also look at, will give
people time to give us feedback when something is off with the
migration we're doing in
https://gitlab.com/gitlab-com/gl-infra/observability/team/-/issues/2573.

Furthermore it helps a bit with
https://gitlab.com/gitlab-com/gl-infra/observability/team/-/issues/4078. As
those panels allow dragging around again.

Signed-off-by: Taliesin Millhouse <tmillhouse@gitlab.com>