fixup: Remove job label from nat port saturation queries
This change removes the `job` matching label from the queries that
make up the nat port saturation metrics. The dashboards using the
metrics are currently not showing metrics due to a change in labels.

We don't require the label for aggregations, and we have the
`gateway_name` label to aggregate on.