Make db_sec*_duration_s log fields float not long
As we discovered when rolling out the new Sec database in
https://gitlab.com/gitlab-com/gl-infra/production/-/issues/19491#note_2414498976
the new log entries for sec DB are defaulting to long which can also be
seen at
https://nonprod-log.gitlab.net/app/management/data/index_management/indices/index_details?indexName=pubsub-rails-inf-gstg-001946&tab=mappings
. Since the first log entry is likely to be 0 it means that
Elasticsearch is assuming that the field a `long` and this causes all
entries to be rounded to the nearest `long`. As can be seen from this
file we already explicitly set the other durations to `float` to ensure
that Elasticsearch ingests them with full precision.