fix: ignore GraphqlController#execute for apdex alerts in reference architectures
This ignores the `GraphqlContoller#execute` from the `puma SLI for
reference architectures and dedicated. We do this because in low
traffic environments, some clients that often trigger requests can
have a disproportionate effect on the apdex and cause a noisy
alert.

At the time of writing, this was caused by an `IntelliJ-GitLab-Plugin`
for which we can't modify the generated GraphQL queries.

This is fine to do, because we do still monitor the Apdex of GraphQL
queries that our own frontend executes in the `graphql_query` SLI.

Related to incident: https://app.incident.io/gitlab/incidents/278