feat(metrics-catalog): add experimental to SLIs
This allows defining an SLI as experimental. An experimental SLI will
generate all of the recording rules, but no alerts.

This allows us to use the analysis dashboards to determine the right
SLO for an SLI without having to silence alerts or reduce the
severity.

This illustrates how this works by marking the `rails_queueing` SLI as
experimental and removing the alerts for that SLI from the GET-hybrid
webservice.

The SLI was initially added to the GET-hybrid in
https://gitlab.com/gitlab-com/runbooks/-/merge_requests/8684/. But
when discussing this on Slack, we decided that we'd rather not alert
on it in the beginning.